Extract InkProject to its own file
